# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "福利签API文档",
        "title": "福利签API文档",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/files/download": {
            "get": {
                "description": "download file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "下载文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "filename",
                        "name": "filename",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "\"avatar\", \"poster\"",
                        "description": "file type",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":true}",
                        "schema": {
                            "$ref": "#/definitions/server.BaseResponse"
                        }
                    }
                }
            }
        },
        "/files/upload": {
            "post": {
                "description": "upload file",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "upload file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "\"avatar\", \"poster\"",
                        "description": "file type",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":true}",
                        "schema": {
                            "$ref": "#/definitions/server.BaseResponse"
                        }
                    }
                }
            }
        },
        "/merchants": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get merchant list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商户"
                ],
                "summary": "获取商户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商户名",
                        "name": "store_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "联系人",
                        "name": "contact_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "联系电话",
                        "name": "contact_phone",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page_no",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "页数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":true}",
                        "schema": {
                            "$ref": "#/definitions/server.MerchantListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "create new merchant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商户"
                ],
                "summary": "新增商户",
                "parameters": [
                    {
                        "description": "参数",
                        "name": "args",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.MerchantVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":true}",
                        "schema": {
                            "$ref": "#/definitions/server.MerchantAddResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Merchant": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "地址",
                    "type": "string"
                },
                "catering_type": {
                    "description": "餐饮类型",
                    "type": "string"
                },
                "checkin_days": {
                    "description": "签到天数多少天可领取礼品",
                    "type": "integer"
                },
                "checkin_num": {
                    "description": "达到指定签到天数后，可领取的礼品数量",
                    "type": "integer"
                },
                "contact_name": {
                    "description": "联系人",
                    "type": "string"
                },
                "contact_phone": {
                    "description": "联系人电话",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "description": "纬度",
                    "type": "number"
                },
                "lon": {
                    "description": "经度",
                    "type": "number"
                },
                "poster": {
                    "description": "商户海报",
                    "type": "string"
                },
                "received": {
                    "description": "已领礼品数量",
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "store_avatar": {
                    "description": "店铺头像",
                    "type": "string"
                },
                "store_name": {
                    "description": "店名",
                    "type": "string"
                },
                "total_receive": {
                    "description": "该店礼品一共可领取总数",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
...
        "model.MerchantVO": {
            "type": "object",
            "required": [
                "address",
                "contact_phone",
                "lat",
                "lon",
                "store_avatar",
                "store_name",
                "total_receive"
            ],
            "properties": {
                "address": {
                    "description": "地址",
                    "type": "string"
                },
                "catering_type": {
                    "description": "餐饮类型",
                    "type": "string"
                },
                "checkin_days": {
                    "description": "签到天数多少天可领取礼品",
                    "type": "integer"
                },
                "checkin_num": {
                    "description": "达到指定签到天数后，可领取的礼品数量",
                    "type": "integer"
                },
                "contact_name": {
                    "description": "联系人",
                    "type": "string"
                },
                "contact_phone": {
                    "description": "联系人电话",
                    "type": "string"
                },
                "lat": {
                    "description": "纬度",
                    "type": "number"
                },
                "lon": {
                    "description": "经度",
                    "type": "number"
                },
                "poster": {
                    "description": "商户海报",
                    "type": "string"
                },
                "received": {
                    "description": "已领礼品数量",
                    "type": "integer"
                },
                "store_avatar": {
                    "description": "店铺头像",
                    "type": "string"
                },
                "store_name": {
                    "description": "店名",
                    "type": "string"
                },
                "total_receive": {
                    "description": "该店礼品一共可领取总数",
                    "type": "integer"
                }
            }
        },
...
        "server.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "server.MerchantAddResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "string"
                },
                "error": {
                    "description": "Error信息",
                    "type": "string"
                },
                "message": {
                    "description": "提示消息",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "boolean"
                }
            }
        },
...
        "server.MerchantListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Merchant"
                    }
                },
                "error": {
                    "description": "Error信息",
                    "type": "string"
                },
                "message": {
                    "description": "提示消息",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}